Added AMD64 impl.
Because I don't know a damn thing about assembly or AMD64, this is just
the output of 6g on my laptop. I just wanted to get a feel for build
directives and what it would be like to have assembly in the build
process.

So if you know what you're doing here, hot diggity send me a pull
request.