Remove the AMD64 impl of the ChaCha20 core.
It really only makes sense if it's faster, and it's not. The 6g compiler
picks up on the bitwise rotation and emits ROLL commands, which was the
only thing I could think of. So until we get a real assembly programmer
in here, I'd rather not pretend this is optimized.