Add support for XChaCha20, the extended nonce variant.
Test vector generated with Andrew Moon's chacha-opt library.  Since the
only difference between the two algorithms is that the initialization
uses a cascade setup (the key/nonce passed into the ctor are passed
through HChaCha to derive a subkey, which then initializes a ChaCha20
instance), this should be reasonably sufficient to test functionality.